Buttons are here!
Added the functionality for buttons to control the maze. All five barriers can be opened/closed, both reward pumps can be manually used and the barriers can be adjusted with buttons. Logic has been implemented to control how the maze handles the door opening when the mouse makes a decison and does not commit, it allows the trial to be completed without the need for interruption.