Added logic to the laser, and the ability to randomize the array of ons.
I added a shuffle function so the array can be randomly "shuffled" to decide which trials will have the laser on or off. The maze does not turn on from the python script. I need to fix that bug but it wo
rks when the input is passed in through serial monitor.